styles: add page size, review margins
weasy print's css supports the @page 'size' attribute, allowing us
to specify a pdf size different from the default A4 size.

The remarkable is 1872 x 1404 pixels, but using that makes fonts too
small without further adjusments so setting a page size of half that is
easier to use.

Having a proper page size, we can re-adjust the margins to better fit
the new format. Also made the space between columns in FifthAvenue
slightly bigger as a matter of taste.

Fixes: #75